Move followable interactions interfce to common package
