+++ models/src/main/java/com/vimeo/networking2/UserInteractions.kt
@@ -1,5 +1,7 @@
 package com.vimeo.networking2
 
+import com.vimeo.networking2.common.FollowableInteractions
+
 /**
  * All actions that can be taken on a user.
  */
