+++ models/src/main/java/com/vimeo/networking2/common/Followable.kt
@@ -1,6 +1,5 @@
 package com.vimeo.networking2.common
 
-import com.vimeo.networking2.FollowableInteractions
 import com.vimeo.networking2.Metadata
 
 /**
