+++ models/src/main/java/com/vimeo/networking2/common/FollowableInteractions.kt
@@ -1,4 +1,6 @@
-package com.vimeo.networking2
+package com.vimeo.networking2.common
+
+import com.vimeo.networking2.UpdatableInteraction
 
 /**
  * Followable interactions.
